DOC-223. Re-read keystore from storage when user-supplied password can't open it
